+++ Changelog.md
@@ -1,14 +1,15 @@
 # Changelog
 
 ## Version (NEXT)
+* added device namespace in domain
 * simplified circuit optimization pipeline code
   * introduced optimization context
+* moved topology namespace to domain
 * enabled topology optimizations to the hardware simulator
 * added statistics for circuits and optimizations
 * made current QASM 2/3 handling more robust
 * added EBNF grammars for OpenQASM 2/3 adapted from the ANTLR grammars
 * added OpenQASM 2/3 parsers with instaparse for improved compatibility including expression evaluation
-* moved topology namespace to domain
 * improved docstrings
 
 ## Version 0.15.0